Fix remaining CodeQL warnings: precision loss, useless assignments, and container access
- Fix precision loss warning in PerformanceMonitoringService (line 200)
- Remove unused exception variables in RealTimeValidationEngine (3 fixes)
- Add container access checks to satisfy CodeQL in:
  * ValidationEngine (_testScenarios, _validationMetrics)
  * SecurityAuditSystem (_vulnerabilities, _incidents, _securityControls)
  * QualityAssuranceSystem (_qualityMetrics, _defects, _qualityControls)
  * AerospaceComplianceSystem (_violations, _certifications)
  * AerospaceReadinessAssessment (_readinessMetrics)
  * QuantumClassicalHybridEngine (_quantumResults, _hybridResults, _materialResults)
  * AutonomousEngineDesigner (_optimizationHistory, _failurePredictions)
  * AIOptimizationEngine (_population, _optimizationHistory)
  * AdvancedErrorHandler (errorContext)
  * AdvancedEngineCore (componentHealth)

Fixes 1 precision loss warning, 3 useless assignment warnings, and 20+ container access warnings

+++ Core/PerformanceMonitoringService.cs
@@ -197,7 +197,8 @@ namespace HB_NLP_Research_Lab.Core
                 RecordMetric("GC_Gen0_Collections", gen0Collections, "Memory");
                 RecordMetric("GC_Gen1_Collections", gen1Collections, "Memory");
                 RecordMetric("GC_Gen2_Collections", gen2Collections, "Memory");
-                RecordMetric("GC_Total_Memory", GC.GetTotalMemory(false) / 1024 / 1024, "Memory"); // MB
+                // Explicit cast to double to avoid precision loss warning (intentional conversion from bytes to MB)
+                RecordMetric("GC_Total_Memory", (double)GC.GetTotalMemory(false) / 1024.0 / 1024.0, "Memory"); // MB
             }
             catch (Exception ex)
             {
